Converts np.longdouble to np.double

+++ stingray/tests/test_io.py
@@ -143,8 +143,8 @@ class TestFileFormats(object):
         assert rec_object['str'] == test_object.str
         assert (rec_object['list'] == test_object.list).all()
         assert (rec_object['array'] == np.array(test_object.array)).all()
-        assert rec_object['long_number'] == test_object.long_number
-        assert (rec_object['long_array'] == np.array(test_object.long_array)).all()
+        assert rec_object['long_number'] == np.double(test_object.long_number)
+        assert (rec_object['long_array'] == np.double(np.array(test_object.long_array))).all()
         os.remove('test.hdf5')
 
     def test_save_ascii(self):
